+++ pyslvs/graph.pyx
@@ -14,8 +14,11 @@ email: pyslvs@gmail.com
 
 cimport cython
 from libcpp.pair cimport pair as cpair
+from numpy cimport uint8_t
 import sys
 from typing import Tuple, Dict, Iterator
+from operator import itemgetter
+from numpy import zeros, uint8
 
 ctypedef cpair[int, int] ipair
 
@@ -190,6 +193,35 @@ cdef class Graph:
     cpdef dict degrees(self):
         """Return number of neighbors par node."""
         return {n: len(neighbors) for n, neighbors in self.adj.items()}
+
+    cpdef ullong degree_code(self):
+        """Return degree code of the graph."""
+        cdef int n = len(self.vertices)
+        if n < 2:
+            return 0
+        cdef imap m
+        cdef int i, n1, n2
+        for i, (n1, _) in enumerate(sorted(
+            self.degrees().items(),
+            key=itemgetter(1),
+            reverse=True
+        )):
+            m[n1] = i
+        # TODO: Available in Cython 0.29.14
+        # cdef bint[:, :] am = zeros((n, n), dtype=bool)
+        cdef uint8_t[:, :] am = zeros((n, n), dtype=uint8)
+        for n1, n2 in self.edges:
+            n1 = m[n1]
+            n2 = m[n2]
+            if n1 > n2:
+                n1, n2 = n2, n1
+            am[n1, n2] = 1
+        cdef ullong code = 0
+        for n1 in range(n):
+            for n2 in range(n1 + 1, n):
+                code <<= 1
+                code += am[n1, n2]
+        return code
 
     cpdef bint is_connected(self, int without = -1):
         """Return True if the graph is not isolated."""